# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Enterprise Recruitment Automation Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job postings
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a job posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job posting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CANDIDATES ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, department, type, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        department:
          type: string
        description:
          type: ["string", "null"]
        competencies:
          type: array
          items:
            type: string
        type:
          type: string
          enum: [internal, external, both]
        status:
          type: string
          enum: [draft, active, closed, expired]
        source:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        requisitionId:
          type: ["integer", "null"]
        applicationCount:
          type: integer
        createdAt:
          type: string
    JobInput:
      type: object
      required: [title, department, type]
      properties:
        title:
          type: string
        department:
          type: string
        description:
          type: ["string", "null"]
        competencies:
          type: array
          items:
            type: string
        type:
          type: string
        expiryDate:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        requisitionId:
          type: ["integer", "null"]
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        type:
          type: string
        expiryDate:
          type: ["string", "null"]
        competencies:
          type: array
          items:
            type: string

    # CANDIDATES
...