# .migration-backup/lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Enterprise Recruitment Automation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface JobInput {
  title: string;
  department: string;
  /** @nullable */
  description?: string | null;
  competencies?: string[];
  type: string;
  /** @nullable */
  expiryDate?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  requisitionId?: number | null;
}